Videos_descriptions_generator/requirements.txt
google-generativeai>=0.3.0
pathlib>=1.0.1
python-dotenv~
